Implement install from ref
Previously, one could only install from an explicit version, such as
v0.4.4 or nightly, which would then download the prebuilt archive from
GitHub and install it.

This is preferable, as you don't need to compile
neovim every time you want to update, but as of this writing, the
nightly builds are paused due to some infrastructure reasons and before
that, the macOS nightly builds would frequently timeout, which would
delete the entire macOS build from the releases page in GitHub, meaning
you couldn't download even the previous nightly build.

As I am impatient, I want the ability to install from any ref that I
want, be it a branch, a tag, or a specific commit.